+++ ci/azure/install-posix.sh
@@ -23,5 +23,6 @@ if [[ -n ${CYTHON} ]]; then CMD="$CMD==${CYTHON}"; fi;
 CMD="$CMD pandas"
 CMD="$CMD $EXTRA"
 if [[ ${USE_CONDA} == "true" ]]; then CMD="$CMD numba"; fi;
+if [[ ${USE_SCIPY} == "true" ]]; then CMD="$CMD scipy"; fi;
 echo $CMD
 eval $CMD